Benchmarks; network arch; perf; correctness

+++ src/spike.rs
@@ -1,5 +1,5 @@
 #[derive(Debug, Clone, Copy)]
 pub struct Spike {
-  pub recvr_id: u64,
+  pub recvr_id: usize,
   pub v: f64
 }
